pushing indices to cassandra
however, if index clause applies, rows can only be retrieved from Cassandra in a
single batch, as there is no good way to get them in a stream. With large
numbers of rows, this can consume much time and memory, and lead to
exceptions. Therefore, it is possible to turn off index push in config file